// apps/api-server/src/interfaces/mongoose.gen.ts
export type UserDeviceDocument = mongoose.Types.Subdocument & {
  ip: string;
  geo: any;
  clientInfo: any;
  token: number;
  _id: mongoose.Types.ObjectId;
  updatedAt?: Date;
  createdAt?: Date;
};